+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v0.07](https://github.com/realityforge/gir/tree/v0.07) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.06...v0.07)