+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Upgrade the `org.realityforge.braincheck` artifact to version `1.13.0`.
+
 ### [v0.07](https://github.com/realityforge/gir/tree/v0.07) (2019-02-04)
 [Full Changelog](https://github.com/realityforge/gir/compare/v0.06...v0.07)