+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Change Log
 
-### Unreleased
+### [v0.08](https://github.com/realityforge/gir/tree/v0.08) (2019-03-25)
+[Full Changelog](https://github.com/realityforge/gir/compare/v0.07...v0.08)
 
 * Upgrade the `org.realityforge.braincheck` artifact to version `1.13.0`.
 * Add `FileUtil.write(...)` methods to ease writing files relative to current directory.